+++ crates/hashing/src/crc32.rs
@@ -1,6 +1,6 @@
-//! CRC32 implementation according to the CRC-32/ISO-HDLC specification.
+//! CRC-32 implementation according to the CRC-32/ISO-HDLC specification.
 //!
-//! The CRC32 algorithm is a widely used checksum algorithm that is used in many
+//! The CRC-32 algorithm is a widely used checksum algorithm that is used in many
 //! network protocols and file formats. The algorithm is based on a polynomial
 //! division of the input data and a predefined polynomial value.
 //!