+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * `Licensee` enumeration with the description of the publisher of the ROM
 * Support for Zippy encoding format for fast compression
+* New hashing crate that includes CRC-32 and CRC-32C implementations
 
 ### Changed