+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-*
+* Issue with the web frontend and `hardReset()` implementation
 
 ## [0.10.4] - 2024-07-16